espeon/pyproject.toml
[tool.poetry]
name = "espeon"
version = "0.1.0"
description = ""
authors = ["Andrew Li <andrewli@u.northwestern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
